+++ frontend/src/pages/inbounds/form/protocols/index.ts
@@ -1,2 +1,4 @@
 export { default as TunFields } from './tun';
 export { default as TunnelFields } from './tunnel';
+export { default as ShadowsocksFields } from './shadowsocks';
+export { default as WireguardFields } from './wireguard';
